+++ spec/integration/resource/api600/c7000/enclosure_group/delete_spec.rb
@@ -1,7 +1,7 @@
 require 'spec_helper'
 
-klass = OneviewSDK::API500::C7000::EnclosureGroup
+klass = OneviewSDK::API600::C7000::EnclosureGroup
 RSpec.describe klass, integration: true, type: DELETE, sequence: rseq(klass) do
-  let(:current_client) { $client_500 }
-  include_examples 'EnclosureGroupDeleteExample', 'integration api500 context', true
+  let(:current_client) { $client_600 }
+  include_examples 'EnclosureGroupDeleteExample', 'integration api600 context', true
 end