+++ spec/integration/resource/api600/synergy/server_profile_template/update_spec.rb
@@ -1,6 +1,6 @@
 require 'spec_helper'
 
-RSpec.describe OneviewSDK::API500::Synergy::ServerProfileTemplate, integration: true, type: UPDATE do
-  let(:current_client) { $client_500_synergy }
-  include_examples 'ServerProfileTemplateUpdateExample', 'integration api500 context'
+RSpec.describe OneviewSDK::API600::Synergy::ServerProfileTemplate, integration: true, type: UPDATE do
+  let(:current_client) { $client_600_synergy }
+  include_examples 'ServerProfileTemplateUpdateExample', 'integration api600 context'
 end