+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Extends support of the SDK to OneView Rest API version 600 (OneView v4.0).
 #### Features supported
 This release adds support to OneView Rest API version 600 for the hardware variants C7000 and Synergy to the already existing features:
    - Enclosure
+   - Enclosure Group
    - Ethernet Network
    - FC Network
    - FCoE Network
@@ -14,12 +15,15 @@ This release adds support to OneView Rest API version 600 for the hardware varia
    - Internal Link Set
    - Logical Enclosure
    - Logical Interconnect
+   - Logical Interconnect Group
    - SAS Interconnect
    - SAS Interconnect Type
    - SAS Logical Interconnect
+   - SAS Logical Interconnect Group
    - Server Hardware
+   - Server Profile Template
    - Volume
-
+  
 ## v5.3.0
 
 #### New Resources: